add roles permissions in ads

+++ src/ads/services/ads.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { ForbiddenException, Injectable, NotFoundException } from "@nestjs/common";
 import { CreateOrUpdateAd } from "../dto/create-or-update-ad.dto";
 import { UserEntity } from "src/users/model/User.entity";
 import { InjectRepository } from "@nestjs/typeorm";
@@ -9,11 +9,14 @@ import { Ad } from "../dto/ad.dto";
 import { Ads } from "../dto/ads.dto";
 import { ExtendedAd } from "../dto/extended-ad.dto";
 import { ImagesService } from "src/files/services/images.service";
+import { Action, CaslAbilityFactory } from "src/auth/services/casl-ability.factory";
+import { ForbiddenError } from "@casl/ability";
 
 @Injectable()
 export class AdsService {
   constructor(
     @InjectRepository(AdEntity) private adsRepository: Repository<AdEntity>,
+    private caslAbilityFactory: CaslAbilityFactory,
     private adsMapper: AdsMapper,
     private imagesService: ImagesService
   ) {}
@@ -51,19 +54,25 @@ export class AdsService {
   }
 
   async getAds(id: number): Promise<ExtendedAd> {
-    const adEntity = await this.getAdByIdForUser(id);
+    const adEntity = await this.getAdById(id);
 
     return this.adsMapper.toExtendedAdDto(adEntity);
   }
 
-  async removeAd(id: number) {
-    const adEntity = await this.getAdByIdForUser(id);
+  async removeAd(id: number, user: UserEntity) {
+    const adEntity = await this.getAdById(id);
+
+    if (!this.isAllowed(Action.Delete, adEntity, user))
+      throw new ForbiddenException("У вас нет прав доступа или объявление вам не принадлежит");
 
     await this.adsRepository.remove(adEntity);
   }
 
-  async updateAds(id: number, dto: CreateOrUpdateAd) {
-    const adEntity = await this.getAdByIdForUser(id);
+  async updateAds(id: number, dto: CreateOrUpdateAd, user: UserEntity) {
+    const adEntity = await this.getAdById(id);
+
+    if (!this.isAllowed(Action.Update, adEntity, user))
+      throw new ForbiddenException("У вас нет прав доступа или объявление вам не принадлежит");
 
     for (const prop in dto) {
       adEntity[prop] = dto[prop];
@@ -74,8 +83,11 @@ export class AdsService {
     return this.adsMapper.toAdDto(adEntity);
   }
 
-  async updateImage(id: number, image: Express.Multer.File) {
-    const adEntity = await this.getAdByIdForUser(id);
+  async updateImage(id: number, image: Express.Multer.File, user: UserEntity) {
+    const adEntity = await this.getAdById(id);
+
+    if (!this.isAllowed(Action.Update, adEntity, user))
+      throw new ForbiddenException("У вас нет прав доступа или объявление вам не принадлежит");
 
     const filename = this.imagesService.createFilename(image);
 
@@ -86,7 +98,7 @@ export class AdsService {
     return image.buffer;
   }
 
-  public async getAdByIdForUser(id: number): Promise<AdEntity> {
+  public async getAdById(id: number): Promise<AdEntity> {
     const adEntity = await this.adsRepository.findOne({
       relations: { author: true },
       where: {
@@ -97,5 +109,11 @@ export class AdsService {
     if (!adEntity) throw new NotFoundException("Объявление не было найдено");
 
     return adEntity;
+  }
+
+  private isAllowed(action: Action, adEntity: AdEntity, user: UserEntity): boolean {
+    const ability = this.caslAbilityFactory.createForUser(user);
+
+    return ability.can(action, adEntity);
   }
 }
